# docs/api/api-docs.yaml
openapi: 3.1.0
info:
  title: TripFlow API
  description: "API documentation for TripFlow, the travel planning application."
  version: v1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- auth_token: []
...
paths:
...
  /api/v1/itineraries/{itineraryId}/share-links:
    get:
      tags:
      - Itinerary Share Links
      summary: Get Active Share Links
      description: Returns active share links for an itinerary. Owner and accepted
        collaborators can view.
      operationId: getActiveShareLinks
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Share links retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareLinkDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareLinkDTO"
        "404":
          description: Itinerary not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareLinkDTO"
      security:
      - auth_token: []
    post:
      tags:
      - Itinerary Share Links
      summary: Generate Share Link
      description: Generates a public read-only share link with fixed expiration.
        Only itinerary owner can generate.
      operationId: generateShareLink
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Share link generated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShareLinkDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShareLinkDTO"
        "404":
          description: Itinerary not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShareLinkDTO"
      security:
      - auth_token: []
...
  /api/v1/share/{token}:
    get:
      tags:
      - Itinerary Share Links
      summary: Get Shared Itinerary
      description: Returns a read-only itinerary by share token. Public endpoint.
      operationId: getSharedItinerary
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Shared itinerary retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
        "403":
          description: Share link expired or revoked
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
        "404":
          description: Share link not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
...
  /api/v1/itineraries/{itineraryId}/share-links/{shareLinkId}:
    delete:
      tags:
      - Itinerary Share Links
      summary: Revoke Share Link
      description: Revokes a share link. Only itinerary owner can revoke.
      operationId: revokeShareLink
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: shareLinkId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Share link revoked
        "403":
          description: Forbidden
        "404":
          description: Itinerary or share link not found
      security:
      - auth_token: []
...
components:
  schemas:
...
    ActivityDTO:
      type: object
      properties:
        activity:
          type: string
        details:
          type: string
        location:
          $ref: "#/components/schemas/LocationDTO"
        time:
          type: string
        duration:
          type: string
    CoordinatesDTO:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...
    ExternalImageDTO:
      type: object
      properties:
        imageUrl:
          type: string
        altDescription:
          type: string
        authorUsername:
          type: string
    ItineraryDayDTO:
      type: object
      properties:
        day:
          type: integer
          format: int32
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityDTO"
    LocationDTO:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        coordinates:
          $ref: "#/components/schemas/CoordinatesDTO"
    ExtendedItineraryResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 200
          minLength: 0
        place:
          type: string
          maxLength: 200
          minLength: 0
        people:
          type: integer
          format: int32
        budget:
          type: number
          format: double
        date:
          type: string
          maxLength: 50
          minLength: 0
        tags:
          type: array
          items:
            type: string
          maxItems: 30
          minItems: 0
        updatedCount:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - DRAFT
          - PLANNED
          - ONGOING
          - COMPLETED
        days:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDayDTO"
        countDays:
          type: integer
          format: int32
        coverImage:
          $ref: "#/components/schemas/ExternalImageDTO"
        permissions:
          $ref: "#/components/schemas/PermissionsDTO"
      required:
      - title
    PermissionsDTO:
      type: object
      properties:
        view:
          type: boolean
        edit:
          type: boolean
        delete:
          type: boolean
...
    ShareLinkDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        token:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        active:
          type: boolean
...
  securitySchemes:
    auth_token:
      type: apiKey
      name: auth_token
      in: cookie